Declare NodeTests
