+++ fs/dev/tf/module.ts
@@ -1,15 +1,15 @@
 import { io } from '../../io/module.ts'
 import { loadModuleMap } from '../module.f.ts'
-import { type RegisterTest, registerModuleMap } from './module.f.ts'
+import { registerModuleMap } from './module.f.ts'
 import * as nodeTest from 'node:test'
 import { fromIo } from '../../io/module.f.ts'
 import { type ToAsyncOperationMap } from '../../types/effects/module.f.ts'
 import { asyncRun } from '../../types/effects/module.ts'
-import { type All } from '../../types/effects/node/module.f.ts'
+import { type All, type Test } from '../../types/effects/node/module.f.ts'
 
-const map: ToAsyncOperationMap<RegisterTest | All> = {
+const map: ToAsyncOperationMap<Test | All> = {
     all: async (...effects) => Promise.all(effects.map(aMap)),
-    registerTest: async (ctx, name, expectFailure, test) =>
+    test: async (ctx, name, expectFailure, test) =>
         ctx.test(name, { expectFailure }, async t => aMap(test(t))),
 }
