+++ fs/dev/tf/test.f.ts
@@ -4,7 +4,7 @@ import { emptyState, type JsModule } from '../../types/effects/node/virtual/modu
 import { virtual } from '../../types/effects/node/virtual/module.f.ts'
 import { assert, assertEq } from '../module.f.ts'
 import {
-    test, defaultReporter, fmtPath, fmtTerm, fmtImport, ghEscape, isInteger, isIdentifier,
+    testAll, defaultReporter, fmtPath, fmtTerm, fmtImport, ghEscape, isInteger, isIdentifier,
     type Reporter, type Path,
     defaultTest,
 } from './module.f.ts'
@@ -38,7 +38,7 @@ const ok1 = (): unknown => ({ result: ['ok', undefined] as const, duration: 1 })
 const run = (dir: Record<string, JsModule>, initCwd = '.'): readonly[readonly Event[], number] => {
     const [reporter, getEvents] = makeReporter()
     const state = { ...emptyState, root: dir }
-    const [, exitCode] = virtual(state)(test(reporter)(options(initCwd)))
+    const [, exitCode] = virtual(state)(testAll(reporter)(options(initCwd)))
     return [getEvents(), exitCode]
 }
 
@@ -47,7 +47,7 @@ const run = (dir: Record<string, JsModule>, initCwd = '.'): readonly[readonly Ev
 const runMain = (dir: Record<string, JsModule>, github = false): readonly[string, string, number] => {
     const state = { ...emptyState, root: dir }
     const opts = options('.', github)
-    const [finalState, exitCode] = virtual(state)(test(defaultReporter(opts))(opts))
+    const [finalState, exitCode] = virtual(state)(testAll(defaultReporter(opts))(opts))
     return [finalState.stdout, finalState.stderr, exitCode]
 }
